Huge commit
Proper destructor functions at many SDL type pointers. Mouse , Keyboard and Player now independent classes instead of Entity type objects. Mouse and keyboard functions reworked based on what type of content are they controlling (controll corresponding with a classical RTS).

+++ 2dgameengine_sdl/ComponentInit.h
@@ -1,5 +1,5 @@
 #pragma once
 
 static struct TileColliderInit {
-	int x, y, width, height;
+	float x, y, width, height;
 };
